+++ CHANGELOG.md
@@ -33,6 +33,11 @@
 	- [Breaking change]: Remove `chacha20poly1305::seal()`, `chacha20poly1305::open()`, `xchacha20poly1305::seal()` `xchacha20poly1305::open()`.
 	- Add structs `ChaCha20Poly1305` and `XChaCha20Poly1305` that offer equivalent `open()` and `seal()` functions from versions prior to `0.18.0`.
 	- Add support for `seal_inplace()` and `open_inplace()` for `ChaCha20Poly1305` and `XChaCha20Poly1305`. These overwrite data directly instead of copying and allow handling the authentication tag separately.
+- [Breaking change] `orion::hazardous::kdf`:
+	- `orion::hazardous::kdf::argon2i::{derive_key, verify}` have been removed.
+	- Functions are now part of `Argon2<Variant, Threading>` struct in `orion::hazardous::kdf::argon2`.
+	- Argon2(i/id) can now use empty salt. This is only true for the `hazardous` layer. 
+- Add support for Argon2id (sequential-mode only).
 - MSRV bumped to `1.87`
 - Add constants for BLAKE2b: `BLAKE2B_MIN_OUTSIZE, BLAKE2B_MAX_OUTSIZE, BLAKE2B_MIN_KEYSIZE, BLAKE2B_MAX_KEYSIZE` making the conditions more discernable.
 - Add support for the following PQ/T HPKE suites: